+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/notifications/SetChange.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Realm Inc.
+ * Copyright 2022 Realm Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
